P-99, P-74 fixed

+++ protected/views/user-lists/list.php
@@ -1,6 +1,7 @@
 <?php
 
 use app\components\Html;
+use prime\models\ar\Setting;
 
 /**
  * @var \yii\web\View $this
@@ -32,14 +33,18 @@ $this->params['subMenu']['items'] = [];
         'dataProvider' => $userListsDataProvider,
         'columns' => [
             'name',
+            [
+                'attribute' => 'user.name',
+                'label' => \Yii::t('app', 'Owner')
+            ],
             'actions' => [
                 'class' => \kartik\grid\ActionColumn::class,
-                'width' => '100px',
-                'template' => '{read} {mail} {update} {remove}',
+                'width' => '120px',
+                'template' => '{read} {mail} {update} {share} {remove}',
                 'buttons' => [
                     'read' => function($url, $model, $key) {
                         return Html::a(
-                            Html::icon('eye-open'),
+                            Html::icon(Setting::get('icons.read')),
                             ['/user-lists/read', 'id' => $model->id]
                         );
                     },
@@ -53,22 +58,50 @@ $this->params['subMenu']['items'] = [];
                             ]
                         );
                     },
+                    'share' => function($url, $model, $key) {
+                        $result = '';
+                        /** @var \prime\models\ar\UserList $model */
+                        if($model->userCan(\prime\models\permissions\Permission::PERMISSION_SHARE)) {
+                            $result = Html::a(
+                                Html::icon(Setting::get('icons.share')),
+                                ['/user-lists/share', 'id' => $model->id],
+                                [
+                                    'title' => \Yii::t('app', 'Share')
+                                ]
+                            );
+                        }
+                        return $result;
+                    },
                     'update' => function($url, $model, $key) {
-                        return Html::a(
-                            Html::icon('pencil'),
-                            ['/user-lists/update', 'id' => $model->id]
-                        );
+                        $result = '';
+                        /** @var \prime\models\ar\UserList $model */
+                        if($model->userCan(\prime\models\permissions\Permission::PERMISSION_WRITE)) {
+                            $result = Html::a(
+                                Html::icon(Setting::get('icons.update')),
+                                ['/user-lists/update', 'id' => $model->id]
+                            );
+                        }
+                        return $result;
                     },
                     'remove' => function($url, $model, $key) {
-                        return Html::a(
-                            Html::icon('trash'),
-                            ['/user-lists/delete', 'id' => $model->id],
-                            [
-                                'class' => 'text-danger',
-                                'data-confirm' => \Yii::t('app', 'Are you sure you want to delete list <strong>{name}</strong>?', ['name' => $model->name]),
-                                'data-method' => 'delete'
-                            ]
-                        );
+                        $result = '';
+                        /** @var \prime\models\ar\UserList $model */
+                        if($model->userCan(\prime\models\permissions\Permission::PERMISSION_SHARE)) {
+                            $result = Html::a(
+                                Html::icon(Setting::get('icons.remove')),
+                                ['/user-lists/delete', 'id' => $model->id],
+                                [
+                                    'class' => 'text-danger',
+                                    'data-confirm' => \Yii::t(
+                                        'app',
+                                        'Are you sure you want to delete list <strong>{name}</strong>?',
+                                        ['name' => $model->name]
+                                    ),
+                                    'data-method' => 'delete'
+                                ]
+                            );
+                        }
+                        return $result;
                     }
 
                 ]
